jwttest/requirements.txt
Django==1.11
djangorestframework==3.6.0
djangorestframework-jwt==1.11.0
Markdown==2.6.11
PyJWT==1.6.1
pytz==2018.3
